+++ puzzle/class_permuter.cc
@@ -18,8 +18,8 @@ void ClassPermuter::iterator::BuildCurrent() {
   }
 }
 
-long long ClassPermuter::permutation_count() const {
-  long long ret = 1;
+double ClassPermuter::permutation_count() const {
+  double ret = 1;
   int value_count = descriptor_->Values().size();
   for (int i = 2; i <= value_count; i++ ) {
     ret *= i;
